apps-contrib - radiobox - even more debugging on the controllerhf.so
