+++ apps-contrib/radiobox/src/Makefile
@@ -8,7 +8,7 @@ CROSS_COMPILE ?= arm-linux-gnueabihf-
 CC=$(CROSS_COMPILE)gcc
 RM=rm
 
-OBJECTS=main.o worker.o cb_ws.o fpga.o calib.o
+OBJECTS=main.o worker.o cb_http.o cb_ws.o fpga.o calib.o
 
 CFLAGS+= -Wall -Werror -g -fPIC $(INCLUDE)
 LDFLAGS=-shared
